Trait webb_event_watcher_traits::evm::EventHandler

pub trait EventHandler {
    type Contract: Deref<Target = Contract<EthersTimeLagClient>> + WatchableContract;
    type Events: EthLogDecode + Clone;
    type Store: HistoryStore + EventHashStore;

    // Required methods
    fn handle_event<'life0, 'life1, 'async_trait>(
        &'life0 self,
        store: Arc<Self::Store>,
        contract: &'life1 Self::Contract,
        __arg3: (Self::Events, LogMeta),
        metrics: Arc<Mutex<Metrics>>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn can_handle_events<'life0, 'life1, 'async_trait>(
        &'life0 self,
        __arg1: (Self::Events, LogMeta),
        wrapper: &'life1 Self::Contract
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

A trait that defines a handler for a specific set of event types.

The handlers are implemented separately from the watchers, so that we can have one event watcher and many event handlers that will run in parallel.

Required Associated Types

type Contract: Deref<Target = Contract<EthersTimeLagClient>> + WatchableContract

The Type of contract this handler is for, Must be the same as the contract type in the watcher.

type Events: EthLogDecode + Clone

The type of event this handler is for.

type Store: HistoryStore + EventHashStore

The storage backend that this handler will use.

Required Methods

fn handle_event<'life0, 'life1, 'async_trait>( &'life0 self, store: Arc<Self::Store>, contract: &'life1 Self::Contract, __arg3: (Self::Events, LogMeta), metrics: Arc<Mutex<Metrics>> ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

a method to be called with the event information, it is up to the handler to decide what to do with the event.

If this method returned an error, the handler will be considered as failed and will be discarded. to have a retry mechanism, use the EventHandlerWithRetry::handle_event_with_retry method which does exactly what it says.

If this method returns Ok(true), the event will be marked as handled.

fn can_handle_events<'life0, 'life1, 'async_trait>( &'life0 self, __arg1: (Self::Events, LogMeta), wrapper: &'life1 Self::Contract ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Whether any of the events could be handled by the handler