Trait webb_event_watcher_traits::evm::EventWatcher

pub trait EventWatcher {
    type Contract: Deref<Target = Contract<Provider<Http>>> + WatchableContract;
    type Events: EthLogDecode + Clone;
    type Store: HistoryStore + EventHashStore;

    const TAG: &'static str;

    fn run<'life0, 'life1, 'async_trait>(
        &'life0 self,
        client: Arc<Provider<Http>>,
        store: Arc<Self::Store>,
        contract: Self::Contract,
        handlers: Vec<EventHandlerFor<Self>>,
        ctx: &'life1 RelayerContext
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
    { ... }
}

A trait for watching events from a watchable contract. EventWatcher trait exists for deployments that are smart-contract / EVM based

Required Associated Types

type Contract: Deref<Target = Contract<Provider<Http>>> + WatchableContract

The contract that this event watcher is watching.

type Events: EthLogDecode + Clone

The Events that this event watcher is interested in.

type Store: HistoryStore + EventHashStore

The Storage backend that will be used to store the required state for this event watcher

Required Associated Constants

const TAG: &'static str

A Helper tag used to identify the event watcher during the logs.

Provided Methods

fn run<'life0, 'life1, 'async_trait>(
&'life0 self,
client: Arc<Provider<Http>>,
store: Arc<Self::Store>,
contract: Self::Contract,
handlers: Vec<EventHandlerFor<Self>>,
ctx: &'life1 RelayerContext
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Returns a task that should be running in the background that will watch events