Trait webb_event_watcher_traits::substrate::SubstrateEventWatcher

pub trait SubstrateEventWatcher {
    type RuntimeConfig: Config + Send + Sync + 'static;
    type Client: OnlineClientT<Self::RuntimeConfig> + Send + Sync;
    type Event: Decode + Send + Sync + 'static;
    type FilteredEvent: StaticEvent + Send + Sync + 'static;
    type Store: HistoryStore;

    const TAG: &'static str;
    const PALLET_NAME: &'static str;

    fn handle_event<'life0, 'async_trait>(
        &'life0 self,
        store: Arc<Self::Store>,
        client: Arc<Self::Client>,
        __arg3: (Self::FilteredEvent, BlockNumberOf<Self>),
        metrics: Arc<Mutex<Metrics>>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;

    fn run<'life0, 'async_trait>(
        &'life0 self,
        chain_id: u32,
        client: Arc<Self::Client>,
        store: Arc<Self::Store>,
        metrics: Arc<Mutex<Metrics>>,
        event_watcher_config: EventsWatcherConfig
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
    { ... }
}

Represents a Substrate event watcher.

Required Associated Types

type RuntimeConfig: Config + Send + Sync + 'static

The Config of this Runtime [subxt::PolkadotConfig, subxt::SubstrateConfig]

type Client: OnlineClientT<Self::RuntimeConfig> + Send + Sync

The Runtime Client that can be used to perform API calls.

type Event: Decode + Send + Sync + 'static

All types of events that are supported by this Runtime. Usually it will be [my_runtime::api::Event] which is an enum of all events.

type FilteredEvent: StaticEvent + Send + Sync + 'static

The kind of event that this watcher is watching.

type Store: HistoryStore

The Storage backend, used by the event watcher to store its state.

Required Associated Constants

const TAG: &'static str

A helper unique tag to help identify the event watcher in the tracing logs.

const PALLET_NAME: &'static str

Pallet name used to fetch pallet id for given target system

Required Methods

fn handle_event<'life0, 'async_trait>(
&'life0 self,
store: Arc<Self::Store>,
client: Arc<Self::Client>,
__arg3: (Self::FilteredEvent, BlockNumberOf<Self>),
metrics: Arc<Mutex<Metrics>>
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

A method to be called with the event information, it is up to the handler to decide what to do with the event. If this method returned an error, the handler will be considered as failed and will be retried again, depends on the retry strategy.

Provided Methods

fn run<'life0, 'async_trait>(
&'life0 self,
chain_id: u32,
client: Arc<Self::Client>,
store: Arc<Self::Store>,
metrics: Arc<Mutex<Metrics>>,
event_watcher_config: EventsWatcherConfig
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,

Returns a task that should be running in the background that will watch events